DOC Defined major sections of the manual.

+++ .travis/install.sh
@@ -20,8 +20,8 @@ else
   QZIP=linuxx86.zip
 fi
 
-if [[ -n ${KDBURL} ]]; then
+if [[ -n ${KDBURL} && ! -d $HOME/d ]]; then
   curl -e https://kx.com "${KDBURL}/3.5/${QZIP}" \
-       -o $HOME/kx.zip && unzip -d $HOME $HOME/kx.zip
+       -o $HOME/d/kx.zip && unzip -d $HOME $HOME/d/kx.zip
   rm $HOME/q/q.q
 fi
